30days-weather-dashboard-main/requirements.txt
boto3==1.26.137
python-dotenv==1.0.0
requests==2.28.2
streamlit
altair==4.2.2

